conversion to namespace step 05
branch created from "Move itemsetup form" (step 4 of the process)
behat tests are 100% working
