+++ format/label/classes/item.php
@@ -15,34 +15,61 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * This file contains the surveyproformat_fieldset
+ * This file contains the surveyproformat_label
  *
- * @package   surveyproformat_fieldset
+ * @package   surveyproformat_label
  * @copyright 2013 onwards kordan <kordan@mclink.it>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-// namespace mod_surveypro;
+namespace surveyproformat_label;
 
 defined('MOODLE_INTERNAL') || die();
 
 use mod_surveypro\itembase;
 
-require_once($CFG->dirroot.'/mod/surveypro/format/fieldset/lib.php');
+require_once($CFG->dirroot.'/mod/surveypro/format/label/lib.php');
 
 /**
- * Class to manage each aspect of the fieldset item
+ * Class to manage each aspect of the label item
  *
- * @package   surveyproformat_fieldset
+ * @package   surveyproformat_label
  * @copyright 2013 onwards kordan <kordan@mclink.it>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class surveyproformat_fieldset_format extends itembase {
+class item extends itembase {
 
     /**
-     * @var string Label of the fieldset
+     * @var string $content
      */
-    protected $content;
+    public $content = '';
+
+    /**
+     * @var int $contenttrust
+     */
+    public $contenttrust = 1;
+
+    /**
+     * @var string $contentformat
+     */
+    public $contentformat = '';
+
+    /**
+     * @var string Custom number of the item
+     *
+     * It usually is 1, 1.1, a, 2.1.a..
+     */
+    protected $customnumber;
+
+    /**
+     * @var int The label has a dedicated row
+     */
+    protected $fullwidth;
+
+    /**
+     * @var string Label on the left of the label content
+     */
+    protected $leftlabel;
 
     /**
      * @var bool Can this item be parent?
@@ -65,10 +92,10 @@ class surveyproformat_fieldset_format extends itembase {
 
         // List of properties set to static values.
         $this->type = SURVEYPRO_TYPEFORMAT;
-        $this->plugin = 'fieldset';
+        $this->plugin = 'label';
 
         // Override the list of fields using format, whether needed.
-        $this->fieldsusingformat = array();
+        // Nothing to override, here.
 
         // Other element specific properties.
         // No properties here.
@@ -77,13 +104,11 @@ class surveyproformat_fieldset_format extends itembase {
         // No properties here.
 
         // List of fields I do not want to have in the item definition form.
-        $this->insetupform['trimonsave'] = false;
-        $this->insetupform['customnumber'] = false;
         $this->insetupform['position'] = false;
+        $this->insetupform['trimonsave'] = false;
         $this->insetupform['extranote'] = false;
         $this->insetupform['required'] = false;
         $this->insetupform['variable'] = false;
-        $this->insetupform['indent'] = false;
         $this->insetupform['hideinstructions'] = false;
 
         if (!empty($itemid)) {
@@ -118,6 +143,7 @@ class surveyproformat_fieldset_format extends itembase {
         // Now execute very specific plugin level actions.
 
         // Begin of: plugin specific settings (eventually overriding general ones).
+        $this->item_custom_fields_to_db($record);
         // End of: plugin specific settings (eventually overriding general ones).
 
         // Do parent item saving stuff here (mod_surveypro_itembase::item_save($record))).
@@ -125,12 +151,26 @@ class surveyproformat_fieldset_format extends itembase {
     }
 
     /**
-     * Get content.
+     * Traslate values from the mform of this item to values for corresponding properties.
      *
-     * @return the content of $content property
+     * @param object $record
+     * @return void
      */
-    public function get_content() {
-        return s($this->content);
+    public function item_custom_fields_to_db($record) {
+        // 1. Special management for composite fields.
+        // Nothing to do: they don't exist in this plugin.
+
+        // 2. Override few values.
+        // Nothing to do: no need to overwrite variables.
+
+        // 3. Set values corresponding to checkboxes.
+        // Take care: 'required', 'trimonsave', 'hideinstructions' were already considered in get_common_settings.
+        $checkboxes = array('fullwidth');
+        foreach ($checkboxes as $checkbox) {
+            $record->{$checkbox} = (isset($record->{$checkbox})) ? 1 : 0;
+        }
+
+        // 4. Other.
     }
 
     /**
@@ -141,7 +181,10 @@ class surveyproformat_fieldset_format extends itembase {
      * @return void
      */
     public function item_add_mandatory_plugin_fields(&$record) {
-        $record->content = 'Fieldset';
+        $record->content = 'Label';
+        $record->contentformat = 1;
+        $record->indent = 0;
+        $record->fullwidth = 0;
     }
 
     // MARK get.
@@ -156,13 +199,26 @@ class surveyproformat_fieldset_format extends itembase {
     }
 
     /**
+     * Get indent.
+     *
+     * @return void
+     */
+    public function get_indent() {
+        if ($this->fullwidth) {
+            return false;
+        } else {
+            return $this->indent;
+        }
+    }
+
+    /**
      * Make the list of the fields using multilang
      *
      * @return array of felds
      */
     public function get_multilang_fields() {
         $fieldlist = array();
-        $fieldlist[$this->plugin] = array('content');
+        $fieldlist[$this->plugin] = array('content', 'leftlabel');
 
         return $fieldlist;
     }
@@ -173,7 +229,7 @@ class surveyproformat_fieldset_format extends itembase {
      * @return the template to use at response report creation
      */
     public static function get_pdf_template() {
-        return 0;
+        return SURVEYPRO_2COLUMNSTEMPLATE;
     }
 
     /**
@@ -185,11 +241,31 @@ class surveyproformat_fieldset_format extends itembase {
         $schema = <<<EOS
 <?xml version="1.0" encoding="UTF-8"?>
 <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" elementFormDefault="qualified">
-    <xs:element name="surveyproformat_fieldset">
+    <xs:element name="surveyproformat_label">
         <xs:complexType>
             <xs:sequence>
                 <xs:element name="content" type="xs:string"/>
-                <xs:element name="defaultstatus" type="xs:int" minOccurs="0"/>
+                <xs:element name="embedded" minOccurs="0" maxOccurs="unbounded">
+                    <xs:complexType>
+                        <xs:sequence>
+                            <xs:element name="filename" type="xs:string"/>
+                            <xs:element name="filecontent" type="xs:base64Binary"/>
+                        </xs:sequence>
+                    </xs:complexType>
+                </xs:element>
+                <xs:element name="contentformat" type="xs:int"/>
+
+                <!-- <xs:element name="required" type="xs:int"/> -->
+                <xs:element name="indent" type="xs:int"/>
+                <!-- <xs:element name="position" type="xs:int"/> -->
+                <xs:element name="customnumber" type="xs:string" minOccurs="0"/>
+                <!-- <xs:element name="hideinstructions" type="xs:int"/> -->
+                <!-- <xs:element name="variable" type="xs:string"/> -->
+                <!-- <xs:element name="extranote" type="xs:string" minOccurs="0"/> -->
+                <!-- <xs:element name="trimonsave" type="xs:int"/> -->
+
+                <xs:element name="fullwidth" type="xs:int"/>
+                <xs:element name="leftlabel" type="xs:string" minOccurs="0"/>
             </xs:sequence>
         </xs:complexType>
     </xs:element>
@@ -212,9 +288,21 @@ EOS;
     public function userform_mform_element($mform, $searchform, $readonly) {
         // This plugin has $this->insetupform['insearchform'] = false; so it will never be part of a search form.
 
-        $mform->addElement('header', $this->itemname, $this->get_content());
-        if ($this->defaultstatus != 2) {
-            $mform->setExpanded($this->itemname, $this->defaultstatus);
+        if ($this->fullwidth) {
+            $content = '';
+            $content .= \html_writer::start_tag('div', array('class' => 'fitem'));
+            $content .= \html_writer::start_tag('div', array('class' => 'fstatic fullwidth label_static'));
+            $content .= $this->get_content();
+            $content .= \html_writer::end_tag('div');
+            $content .= \html_writer::end_tag('div');
+            $mform->addElement('html', $content);
+        } else {
+            $labelsep = get_string('labelsep', 'langconfig'); // Separator usually is ': '.
+            $elementnumber = $this->customnumber ? $this->customnumber.$labelsep : '';
+            $elementlabel = $elementnumber.$this->leftlabel;
+            $attributes = array();
+            $attributes['class'] = 'indent-'.$this->indent.' label_static';
+            $mform->addElement('mod_surveypro_label', $this->itemname, $elementlabel, $this->get_content(), $attributes);
         }
     }
 
@@ -260,7 +348,10 @@ EOS;
      */
     public function userform_get_root_elements_name() {
         $elementnames = array();
-        $elementnames[] = $this->itemname;
+
+        if (empty($this->fullwidth)) {
+            $elementnames[] = $this->itemname;
+        }
 
         return $elementnames;
     }
